v0.8.2.1 update

+++ Projectiles/Weapon/Melee/GenesisRagnarok/RagnarokProj.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using ShardsOfAtheria.Items.Weapons.Melee;
+using ShardsOfAtheria.Projectiles.Weapon.Melee.GenesisRagnarok.IceStuff;
 using System;
 using Terraria;
 using Terraria.Audio;
@@ -26,11 +27,27 @@ namespace ShardsOfAtheria.Projectiles.Weapon.Melee.GenesisRagnarok
             DrawOriginOffsetY = -4;
         }
 
+        public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
+        {
+            Player player = Main.player[Projectile.owner];
+
+            if (player.HeldItem.type == ModContent.ItemType<GenesisAndRagnarok>())
+            {
+                if ((player.HeldItem.ModItem as GenesisAndRagnarok).upgrades < 5 && (player.HeldItem.ModItem as GenesisAndRagnarok).upgrades >= 3)
+                    target.AddBuff(BuffID.OnFire, 600);
+                else if ((player.HeldItem.ModItem as GenesisAndRagnarok).upgrades == 5)
+                {
+                    target.AddBuff(BuffID.Frostburn, 600);
+                    Projectile.NewProjectile(Projectile.GetSource_FromThis(), Projectile.Center, Vector2.Zero, ModContent.ProjectileType<IceVortexShard>(), Projectile.damage, Projectile.knockBack, player.whoAmI);
+                }
+            }
+        }
+
         public override void AI()
         {
             Player player = Main.player[Projectile.owner];
-            if (player.HeldItem.type == ModContent.ItemType<GenesisAndRagnarok>())
-                (player.HeldItem.ModItem as GenesisAndRagnarok).combo = 0;
+            player.itemAnimation = 10;
+            player.itemTime = 10;
 
             int newDirection = Projectile.Center.X > player.Center.X ? 1 : -1;
             player.ChangeDir(newDirection);
@@ -40,32 +57,36 @@ namespace ShardsOfAtheria.Projectiles.Weapon.Melee.GenesisRagnarok
             if (Projectile.ai[1] >= 20)
                 Projectile.velocity = Vector2.Normalize(player.Center - Projectile.Center) *30;
 
-            for (int num72 = 0; num72 < 2; num72++)
+            if (player.HeldItem.type == ModContent.ItemType<GenesisAndRagnarok>())
             {
-                Dust obj4 = Main.dust[Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, DustID.Torch, 0f, 0f, 100, default, 2f)];
-                obj4.noGravity = true;
-                obj4.velocity *= 2f;
-                obj4.velocity += Projectile.localAI[0].ToRotationVector2();
-                obj4.fadeIn = 1.5f;
-            }
+                (player.HeldItem.ModItem as GenesisAndRagnarok).combo = 0;
 
-            if (Projectile.getRect().Intersects(player.getRect()) && Projectile.ai[1] > 20 || player.dead || Main.mouseLeft)
-                Projectile.Kill();
+                for (int num72 = 0; num72 < 2; num72++)
+                {
+                    Dust obj4 = Main.dust[Dust.NewDust(Projectile.position, Projectile.width, Projectile.height, (player.HeldItem.ModItem as GenesisAndRagnarok).upgrades < 5 ? DustID.Torch : DustID.Frost, 0f, 0f, 100, default,
+                        (player.HeldItem.ModItem as GenesisAndRagnarok).upgrades < 5 ? 2f : .5f)];
+                    obj4.noGravity = true;
+                    obj4.velocity *= 2f;
+                    obj4.velocity += Projectile.localAI[0].ToRotationVector2();
+                    obj4.fadeIn = 1.5f;
+                }
+
+                if (Projectile.getRect().Intersects(player.getRect()) && Projectile.ai[1] > 20 || player.dead || (Main.mouseLeft && (Main.LocalPlayer.HeldItem.ModItem as GenesisAndRagnarok).upgrades >= 2))
+                    Projectile.Kill();
+            }
         }
 
         public override void Kill(int timeLeft)
         {
             Player player = Main.player[Projectile.owner];
-            if (!player.dead && Main.mouseLeft && Main.myPlayer == Projectile.owner && player.ownedProjectileCounts[ModContent.ProjectileType<RagnarokProj2>()] == 0)
+            if (player.HeldItem.type == ModContent.ItemType<GenesisAndRagnarok>())
             {
-                Projectile.NewProjectile(Projectile.GetSource_FromThis(), player.Center, Vector2.Zero, ModContent.ProjectileType<RagnarokProj2>(), player.HeldItem.damage, player.HeldItem.knockBack, player.whoAmI);
-                SoundEngine.PlaySound(SoundID.Item71);
+                if (!player.dead && Main.mouseLeft && Main.myPlayer == Projectile.owner && player.ownedProjectileCounts[ModContent.ProjectileType<RagnarokProj2>()] == 0 && (Main.LocalPlayer.HeldItem.ModItem as GenesisAndRagnarok).upgrades >= 2)
+                {
+                    Projectile.NewProjectile(Projectile.GetSource_FromThis(), player.Center, Vector2.Zero, ModContent.ProjectileType<RagnarokProj2>(), Projectile.damage, Projectile.knockBack, player.whoAmI);
+                    SoundEngine.PlaySound(SoundID.Item71);
+                }
             }
-        }
-
-        public override void OnHitNPC(NPC target, int damage, float knockback, bool crit)
-        {
-            target.AddBuff(BuffID.OnFire, 600);
         }
 
         public override Color? GetAlpha(Color lightColor)
